requirements.txt
natsort==7.1.0
unicode_slugify==0.1.3
